pyproject.toml
[tool.poetry]
name = "ezthrottle"
version = "1.0.0"
description = "Python SDK for EZThrottle - The API Dam for rate-limited services"
authors = ["Rahmi Pruitt <rjpruitt16@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ezthrottle"}]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.0"
flask = "^2.0.0"
python-dotenv = "0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
